clean up: create audio recording hook
